Split `PLL_Install` and `PLL_Install_Base` into several classes
